+++ app/tests/TestCase.php
@@ -10,6 +10,27 @@ abstract class TestCase extends BaseTestCase
     public function createApplication()
     {
         $basePath = dirname(__DIR__);
+        $testingEnvironment = [
+            'APP_ENV' => 'testing',
+            'APP_MAINTENANCE_DRIVER' => 'file',
+            'BCRYPT_ROUNDS' => '4',
+            'BROADCAST_CONNECTION' => 'null',
+            'CACHE_STORE' => 'array',
+            'DB_CONNECTION' => 'sqlite',
+            'DB_DATABASE' => ':memory:',
+            'MAIL_MAILER' => 'array',
+            'QUEUE_CONNECTION' => 'sync',
+            'SESSION_DRIVER' => 'array',
+            'PULSE_ENABLED' => 'false',
+            'TELESCOPE_ENABLED' => 'false',
+            'NIGHTWATCH_ENABLED' => 'false',
+        ];
+
+        foreach ($testingEnvironment as $key => $value) {
+            putenv($key.'='.$value);
+            $_ENV[$key] = $value;
+            $_SERVER[$key] = $value;
+        }
 
         $_ENV['APP_BASE_PATH'] = $basePath;
         $_SERVER['APP_BASE_PATH'] = $basePath;